// apps/web/src/lib/api/generated/types/groupMovieResponseDto.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Movies App API
 * REST API for Movies application

**Links:**
- [OpenAPI JSON](/api/docs/json)
- [OpenAPI YAML](/api/docs/yaml)
 * OpenAPI spec version: 1.0.0
 */
import type { GroupMovieResponseDtoStatus } from './groupMovieResponseDtoStatus';

export interface GroupMovieResponseDto {
	id: number;
	groupId: number;
	movieId: number;
	addedBy: number;
	status: GroupMovieResponseDtoStatus;
	plannedDate?: string;
	watchedDate?: string;
	createdAt: string;
	updatedAt: string;
}